requirements.txt
pdfplumber
chromadb
pydantic-settings
fastapi
uvicorn[standard]
twilio
sqlalchemy
psycopg2-binary
google-cloud-aiplatform
python-jose[cryptography]
passlib
pydantic[email]
python-multipart
pdfplumber
alembic
pytz
